FINAL state after development in academic year 2015/16
Implementation contains:
- Jurik's AR for Vuzix glasses and web camera
- Garaj's git repository visualization
- Horvath's Java visualization via city metaphore
- Kostan's projected AR
- team project's 3D mouse, multi-projection, 3D stereo

+++ include/Clustering/Figures/Sphere.h
@@ -14,7 +14,7 @@ class Sphere
 
 public:
 
-	Sphere( osg::Vec3d position, float radius, osg::Vec4d color );
+	Sphere( osg::Vec3d position, double radius, osg::Vec4d color );
 
 	osg::AutoTransform* getAT()
 	{
