+++ include/Data/DbEdge.h
@@ -8,22 +8,22 @@ namespace Data {
 class DbEdge : public VizEdge
 {
 public:
-	DbEdge( qlonglong id, QString name, Data::Graph* graph, bool isOriented, Data::Type* type, float scaling );
+	DbEdge( qlonglong id, QString name, Data::Graph* graph, bool isOriented, Data::Type* type, double scaling );
 
 	/**
-	    *  \fn inline public  isInDB
-	    *  \brief Returns true, if the Edge is already in database
-	    *  \return bool
-	    */
+		*  \fn inline public  isInDB
+		*  \brief Returns true, if the Edge is already in database
+		*  \return bool
+		*/
 	bool isInDB()
 	{
 		return inDB;
 	}
 
 	/**
-	    *  \fn inline public  setIsInDB
-	    *  \brief Sets the inDB flag of the Edge to true (meaning that the Edge is in database)
-	    */
+		*  \fn inline public  setIsInDB
+		*  \brief Sets the inDB flag of the Edge to true (meaning that the Edge is in database)
+		*/
 	void setIsInDB()
 	{
 		inDB = true;
@@ -31,9 +31,9 @@ public:
 
 protected:
 	/**
-	    *  bool inDB
-	    *  \brief Flag if the Edge is already added to database or not
-	    */
+		*  bool inDB
+		*  \brief Flag if the Edge is already added to database or not
+		*/
 	bool inDB;
 
 };
